+++ website_seo_blog/models/website_blog.py
@@ -19,50 +19,8 @@
 #
 ##############################################################################
 from openerp import api, fields, models
-from openerp.addons.web.http import request
-from openerp.addons.website_seo.models import website
-from openerp.addons.website_seo.models.website import slug
+from openerp.addons.website_seo.models.website import slug, META_ROBOTS
 from openerp.tools.translate import _
-
-
-website_url_for_lang = website.url_for_lang
-
-
-def url_for_lang(location, lang):
-    translated_location = website_url_for_lang(location, lang)
-    if translated_location == location:
-        blog_obj = request.registry['blog.blog']
-        ctx = request.context.copy()
-        url_parts = location.split('/')
-        blog_url = url_parts.pop(0)
-        while blog_url in ['', 'blog']:
-            blog_url = url_parts.pop(0)
-        blogs = blog_obj.search(request.cr, request.uid,
-                                [('seo_url', '=', blog_url)],
-                                context=ctx)
-        if blogs and url_parts:
-            post_obj = request.registry['blog.post']
-            posts = post_obj.search(request.cr, request.uid,
-                                    [('blog_id', '=', blogs[0]),
-                                     ('seo_url', '=', url_parts[0])],
-                                    context=ctx)
-            if posts:
-                ctx.update({'lang': lang})
-                post = post_obj.browse(request.cr, request.uid,
-                                       posts[0], context=ctx)
-                location = '/%s/%s' % (post.blog_id.seo_url, post.seo_url)
-        elif blogs:
-            ctx.update({'lang': lang})
-            blog = blog_obj.browse(request.cr, request.uid,
-                                   blogs[0], context=ctx)
-            location = '/%s' % blog.seo_url
-    else:
-        location = translated_location
-    return location
-
-
-# change method url_for_lang to use the one redefined here
-setattr(website, 'url_for_lang', url_for_lang)
 
 
 class Blog(models.Model):
@@ -147,3 +105,52 @@ class BlogPost(models.Model):
     def onchange_name(self, name=False, seo_url=False):
         """If SEO url is empty generate the SEO url when changing the name."""
         return self.env['blog.blog'].onchange_name(name, seo_url)
+
+
+class BlogTag(models.Model):
+
+    """Add SEO url handling for blog posts.
+
+    If you create or update a blog post and this field is empty it
+    is filled automatically when you enter a blog post name.\nIf you fill
+    out this field manually the allowed characters are a-z, A-Z, 0-9, -
+    and _.
+
+    """
+
+    _name = 'blog.tag'
+    _inherit = ['blog.tag', 'website.seo.metadata']
+
+    _sql_constraints = [
+        ('seo_url_uniq', 'unique(seo_url)', _('SEO url must be unique!'))
+    ]
+
+    @api.model
+    def create(self, vals):
+        """Add check for correct SEO urls.
+        """
+        if vals.get('name', False) and not vals.get('seo_url', False):
+            vals['seo_url'] = slug((1, vals['name']))
+
+        if not vals.get('website_meta_robots', False):
+            vals['website_meta_robots'] = self.get_default_meta_robots()
+
+        return super(BlogTag, self).create(vals)
+
+    @api.multi
+    def onchange_name(self, name=False, seo_url=False):
+        """If SEO url is empty generate the SEO url when changing the name."""
+        return self.env['blog.tag'].onchange_name(name, seo_url)
+
+    @api.model
+    def get_default_meta_robots(self):
+        cfg = self.env['website.config.settings'].search([])
+        return cfg and cfg[0].website_blog_tag_default_meta_robots or False
+
+
+class Website(models.Model):
+    _inherit = 'website'
+
+    website_blog_tag_default_meta_robots = fields.Selection(META_ROBOTS,
+                                                            string='Website blog tag default meta robots',
+                                                            translate=True)
