+++ website_seo_blog/views/res_config.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+  <data>
+
+    <record id="view_website_config_settings" model="ir.ui.view">
+      <field name="name">Website Blog settings</field>
+      <field name="model">website.config.settings</field>
+      <field name="inherit_id" ref="website_seo.view_website_config_settings"/>
+      <field name="arch" type="xml">
+        <xpath expr="//form/div" position="inside">
+          <group string="Blogs">
+            <label for="website_blog_tag_default_meta_robots"/>
+            <div name="website_blog_tag_default_meta_robots">
+                <div class="oe_inline">
+                    <field name="website_blog_tag_default_meta_robots" />
+                </div>
+            </div>
+          </group>
+        </xpath>
+      </field>
+    </record>
+
+  </data>
+</openerp>
